% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monday.R
\name{monday}
\alias{monday}
\title{Monday}
\usage{
monday(date)
}
\arguments{
\item{date}{\code{str / date} Date for which to find monday.}
}
\description{
A convenience wrapper to get the monday associated with a particular date.
}
\examples{
monday('2023-06-25')

}
