% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_from_dictionary.R
\name{rename_from_dictionary}
\alias{rename_from_dictionary}
\title{Rename columns of a dataframe using a dictionary}
\usage{
rename_from_dictionary(df, dictionary = NULL, dictionary_fn = NULL, ...)
}
\arguments{
\item{df}{\code{dataframe} Dataframe to be renamed}

\item{dictionary}{\code{char} Named character vector where the names are the new column names and the
values are the old ones.}

\item{dictionary_fn}{\code{char} String indicating the location of a file containing the renaming
dictionary as a dataframe (for exmaple in an csv or xlsx). This dataframe is expected to have
columns called "original_name" and "new_name", if it does not then the first two columns will
be used as "original" and "new" respectively. If \code{dictionary} is provided, this argument will
be ignored.}

\item{...}{Additional arguments to be passed to \code{\link[rio:import]{rio::import()}} when loading \code{dictionary_fn}.
This could be useful, for example  if you have stored the renaming dictionaries for multiple
datasets in a single xlsx where each sheet coresponds to a separate dictionary.}
}
\description{
Uses a dictionary either from a file or named list to rename the relevant columns of
a dataframe.
}
\examples{
df <- data.frame(fizz = c(1, 2, 3),
                 buzz = c('one', 'two', 'three'))

dictionary <- c(foo = 'fizz',
                bar = 'buzz')

rename_from_dictionary(df, dictionary)

}
