% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_columns.R
\name{validate_columns}
\alias{validate_columns}
\title{Validate requried columns}
\usage{
validate_columns(df, cols, select = TRUE)
}
\arguments{
\item{df}{\code{dataframe} Dataframe to be validated.}

\item{cols}{\code{list} List of columns that must be present. If \code{cols} is named, any columns in the
items 'date', 'numeric', and 'character' will have their type validated. In the case of a
named vector, anything that is not meant to be 'date', 'numeric', or 'character' can be placed
in an element named 'other' (or something else, the name is arbitrary).}

\item{select}{\code{bool} If TRUE, only the columns specified in \code{cols} will be returned. All columns
will be returned otherwise.}
}
\description{
Function that checks if all required coluns are present in a df and creates any that
that are missing. Optionally, you can also use a named list for the column specifications to
indicate which ones are date, numeric, string, or other.
}
\examples{
df <- data.frame(date = c('2022-01-01', '2022-02-01'), cases = c('152', '296'), deaths = c(1, 4))
cols <- list(date = 'date',  numeric = c('cases', 'deaths'), other = c('test_result'))
validate_columns(df, cols)

}
